Fix import packages issue

+++ code/perceptron/perceptron.py
@@ -6,5 +6,5 @@ class Perceptron(object):
         self.w2 = w2
         self.b = b
 
-    def output(self, x1: int, x2: int):
+    def output(self, x1: float, x2: float):
         return self.w1*x1 + self.w2*x2 + self.b > 0
